+++ CHANGELOG.md
@@ -8,6 +8,7 @@
 
 - Add a `skip_superimposed` flag to `segment.flattened()` to allow the user to flatten segments without flattening superimposed elements (see #664) (@roussel-ryan)
 - `TransverseDeflectingCavity` now supports `linear` as a `tracking_method`. (see #678) (@cr-xu)
+- Add vectorized histogramming (see #680) (@jp-ga)
 
 ### 🐛 Bug fixes